requirements.txt
fastapi
uvicorn
python-multipart

langchain
langchain_pinecone
langchain_google_genai
langchain_core
langchain_community
pydantic
pypdf

duckduckgo-search

requests
python-dotenv

streamlit